requirements.txt
requests
python-dotenv
dateparser
streamlit
